requirements.txt
Flask
flask-httpauth
scikit-learn
mysql-connector-python
nltk
